requirements.txt
PyQt6==6.5.2
PyMuPDF==1.23.3
Pillow==10.0.0
